requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-storage==0.32.0
middleware==1.2
pymongo==3.3.0
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
